auslagerung getContext
